creating sorting algorithm for strings

+++ ft_ls.h
@@ -6,7 +6,7 @@
 /*   By: efriedma <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/07/24 23:05:55 by efriedma          #+#    #+#             */
-/*   Updated: 2018/07/27 00:07:32 by efriedma         ###   ########.fr       */
+/*   Updated: 2018/11/15 22:28:27 by efriedma         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,7 +27,7 @@
 # define t 16
 
 const unsigned int		g_opt[256];
-unsigned int			g_index;
+short					g_index;
 unsigned int			g_flags;
 
 typedef struct			s_file
@@ -38,6 +38,7 @@ typedef struct			s_file
 	char				*name;
 	//stat puts out permissions and ownership of a directory
 	char				*statformat;
+
 }						t_file;
 
 int						get_opt(int argc, char **argv);
